Helper functions

UltraFast.IOConstructor — Type

struct IOConstructor

A struct representing an IO constructor.

Fields

lookup::Dict{String, Dict{String, Function}}: A dictionary that maps strings to dictionaries of strings and functions. The first level describes the top level abstract type, and the second level corresponds to the name of the object and the constructor function.

UltraFast.extend! — Method

extend(ioconstructor::IOConstructor, category::Symbol, name::String, constructor::Function)

Extend the ioconstructor by adding a new constructor function for a given category and name.

Arguments

ioconstructor::IOConstructor: The IOConstructor object to extend.

category::Symbol: The category of the constructor.

name::String: The name of the constructor.

constructor::Function: The constructor function to add.

UltraFast.relative_error — Method

Calculates the relative error between an estimated energy and the lowest reference energy of a Hamiltonian.

Returns a tuple containing the relative error and a dictionary with information about the reference energy (method and energy).

UltraFast.relative_error — Method

Calculates the relative error between an estimated energy and a reference energy.

UltraFast.vscore — Method

Calculates the v-score for a given system.

UltraFast.FluxComplexInitializers — Module

helper functions for initializing neural networks

UltraFast.FluxComplexInitializers.complex_glorot_normal — Method

complex version of glorot_normal

UltraFast.FluxComplexInitializers.complex_glorot_uniform — Method

complex version of glorot_uniform

UltraFast.FluxComplexInitializers.default_uniform — Method

Default initialization

UltraFast.FluxComplexInitializers.get_number_of_parameters — Method

get the number of parameters of a model

UltraFast.FluxComplexInitializers.imaginary_default_uniform — Method

Default Imaginary initialization

UltraFast.FluxComplexInitializers.real_default_uniform — Method

Default real initialization

UltraFast.Helpers.condense_namedtuples_into_dict — Method

condense_namedtuples_into_dict(vector::AbstractVector, names::Tuple)

Condense a vector of named tuples into a dictionary of arrays, where each key corresponds to a field in the named tuples and each value is an array containing the concatenated values of that field across all named tuples.

Examples

julia> data = [(a=ones(2), b=ones(3)*2) for _ in 1:5];

julia> result = UltraFast.Helpers.condense_namedtuples_into_dict(data)
Dict{Symbol, Any} with 2 entries:
  :a => [1.0 1.0 … 1.0 1.0; 1.0 1.0 … 1.0 1.0]
  :b => [2.0 2.0 … 2.0 2.0; 2.0 2.0 … 2.0 2.0; 2.0 2.0 … 2.0 2.0]

UltraFast.Helpers.flatten_batch — Method

flatten_batch(A; keep=2)

Reshape an array A by flattening all dimensions after the keep-th dimension into a single dimension.

UltraFast.Helpers.random_spin_state — Method

random_spin_state(nspins::Int, mag0::Bool, nsamples::Int)

Generate multiple random spin state for a Heisenberg model.

Arguments

nspins::Int: The number of spins in the system.

mag0::Bool: If true, the total magnetization is set to zero.

nsamples::Int: The number of random samples to generate.

Returns

List of nsamples random spin states for the specified number of spins and samples.

UltraFast.Helpers.random_spin_state — Method

random_spin_state(nspins::Int, mag0::Bool)

Initializes a random spin state with nspins size.

Arguments

nspins::Int: The number of spins in the state.

mag0::Bool: If true, the initial state is generated with zero total magnetization.

Returns

A random spin state with the specified number of spins and magnetization condition.

UltraFast.Hamiltonians.reference_energy — Method

reference_energy(H::Hamiltonian)

Returns a dictionary containing literature reference ground state energy for the given Hamiltonian `H`.