Optimisation

UltraFast.Optimisation.GroundStateOptimisation — Type

struct GroundStateOptimisation

A struct representing the parameters for ground state optimization.

Fields

H::Hamiltonian: The Hamiltonian.

optimiser::Optimiser.AbstractRule: The optimizer. Default is Optimiser.Descent(0.005).

sampler::Sampler: The sampler.

gradient::Gradients: The gradient. Default is SmatrixGradient().

callback::Function: The callback function. Default is () -> nothing.

UltraFast.Optimisation.GroundStateOptimizationHyperParameters — Type

struct GroundStateOptimizationHyperParameters

A structure type containing hyperparameters for ground state optimization.

Fields:

samples::Int64: The number of Monte Carlo samples.

iterations::Int64: The number of iteration steps.

sweep::Int64: The number of flips per sweep. After one sweep the configuration is considered independent.

thermalizationsweeps::Int64: The number of thermalization sweeps expressed as a multiple of the sweeps

UltraFast.Optimisation.windowed_energy — Method

windowed_energy(energy, window_size, at)

Compute the mean energy within a window of a given size. The mean energy is computed as the average of [at - window_size, at].

Arguments

energy: An array of energy values.

window_size: The size of the window.

at: The index at which to compute the windowed energy.

Returns

The mean energy within the specified window.