Optimisation

Samplers

What should a sampler return?

states, the sampled states

logwavefunctions, the log wavefunction values of the sampled states

n_accepted, if MCMC based, the number of accepted moves

n_flips, if MCMC based, the number of proposed moves

n_chains, if parallel chains, the number of chains

runtime, the time it took to do the sampling

UltraFast.Samplers.MCMCSettings — Type

struct MCMCSettings

The MCMCSettings struct represents the settings for a Markov Chain Monte Carlo simulation.

Fields

sweep::Int: The number of sweeps to perform in the simulation.

nsamples::Int: The number of samples to collect during the simulation.

nthermalization::Int: The number of thermalization steps to perform before collecting samples measured in sweeps.

UltraFast.Samplers.generate_flip — Method

generate_flip(state::AbstractArray{<:Integer, 2}, flip_generator::Function)

Generate flips and acceptance flags using a flip generator function.

Arguments

state::AbstractArray{<:Integer, 2}: The state array, or state::AbstractArray{<:Integer, 1}: The state array.

flip_generator::Function: The flip generator function.

Returns

accepts::Array{Bool, 1}: An array of acceptance flags.

flips::Array{Int64, 2}: An array of flips.

Examples

nspins = 4
nparallel = 2

states = UltraFast.GenRandomStates(true, 4, 2)

accepts, flips = UltraFast.generate_flip(states, UltraFast.antiferromagnetic_flip_original)