Translation Invariance

UltraFast.SymmetryRBM.FastTranslationInvariantRBM — Method

FastTranslationInvariantRBM(lattice::LatticeSettings, alpha::Int; init = Models.default_uniform())

Translation-invariant Restricted Boltzmann Machine (RBM) which uses a lookup table for efficient computation of the wavefunction and its derivatives.

Arguments

lattice::LatticeSettings: The lattice configuration.

alpha::Int: The number of hidden units in the RBM.

init: (Optional) Initialization function for RBM parameters. Defaults to Models.default_uniform().

Returns

FastTranslationInvariantRBM{T}: A translation-invariant RBM model where T is the element type of the weight matrix.

Details

The function initializes:

RBM parameters (weights and biases)

A lookup table for efficient computation of the wavefunction

Temporary storage arrays for intermediate calculations

UltraFast.SymmetryRBM.RBMIndepLayer — Type

RBMIndepLayer

A struct representing an independent layer of a Restricted Boltzmann Machine (RBM) neural network.

Fields

Lx::Int: The width of the layer.

Ly::Int: The height of the layer.

alpha::Int: The number of hidden units in the layer.

nspins::Int: The total number of spins in the layer (equal to Lx * Ly).

independent_weight::Array{ComplexType,1}: The independent weights of the layer.

W::Array{ComplexType, 2}: The weight matrix connecting the visible and hidden units.

b::Array{ComplexType, 1}: The bias vector of the hidden units.

Usage

layer = RBMIndepLayer(Lx=2, Ly=2, alpha=2)

Description

The only parameter that should be changed is independent_weight. The other are constants, including W and b.

UltraFast.SymmetryRBM.TranslationInvariantRBM — Method

RBMModelLogIndep(nspins, alpha; initializer)

Translation-invariant Restricted Boltzmann Machine (RBM) model using Flux.jl.

Arguments

nspins::Int: The number of visible units (or spins) in the RBM.

alpha::Int: The number of hidden units in the RBM as a multiple of the number of visible units/nspins.

Returns

A Flux.jl model representing a Restricted Boltzmann Machine with the specified number of visible and hidden units.

Description

This function creates an RBM model using the Flux.jl library.

The complex_glorot_uniform initializer is used to initialize the weights. This initializer is designed to handle complex-valued weights.

Examples

# Create an RBM with 10 visible units and 5 hidden units
model = RBMModel(10, 5)

# Generate a random input vector
input = rand(10)

# Compute the output of the RBM
output = model(input)

UltraFast.SymmetryRBM.TranslationInvarianceTools.WeightTranslator — Type

WeightTranslator

A struct representing a weight translator used for restricted boltzmann machine with periodic boundary conditions.

Fields

independent_weights::Array{ComplexType,1}: An array of independent weights.

weights::Array{ComplexType,2}: A 2D array of weights.

biases::Array{ComplexType,1}: An array of biases.

construction_table_weights::Array{Int64,2}: A 2D array representing the construction table.

Description

The construction_table_weights is a 2D array of index of the independent weights. The first index is the row of the weight matrix. The second index is column of the weight matrix.

UltraFast.SymmetryRBM.TranslationInvarianceTools.WeightTranslator — Method

WeightTranslator(nspins::Int, alpha::Int, Lx::Int, Ly::Int)

Constructs a WeightTranslator object that translates the independent weights into the full weight matrix and biases.

Arguments

nspins::Int: the number of spins in the system

alpha::Int: the number of independent weights

Lx::Int: the number of spins in the x-direction

Ly::Int: the number of spins in the y-direction

Returns

WeightTranslator: a WeightTranslator object that can be used to translate the independent weights into the full weight matrix and biases.

Examples

# Create a weight translator
Lx = 2
Ly = 2
nspins = Lx * Ly
alpha = 2
translator = WeightTranslator(nspins, alpha, Lx, Ly)

# Create a random set of independent weights
independent_weights = randn(ComplexType, alpha + alpha*nspins)

# Translate the independent weights into the full weight matrix and biases
weights, biases = weights_from(translator, independent_weights)